+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Next
 
+* Remove partner ID from user agent string
 * Fixes environment not being used when using MS Graph [[GH-87](https://github.com/hashicorp/vault-plugin-secrets-azure/pull/87)]